refactor(api): split auth/mod.rs into per-concern submodules (#477)
auth/mod.rs held 1042 lines of implementation. Move it into seven named
siblings — types (User + request/response structs), migrate (schema +
role seeding), users (CRUD + audit + lookups), totp_store, api_keys,
revocation, middleware (auth_middleware, AuthUser, require_perm/admin,
perm_check!) — leaving mod.rs as declarations and re-exports. Pure
relocation; the `pub use <mod>::*` re-exports keep every flat
`auth::<name>` path resolving, so no call sites change.

Code verified line-for-line identical (users.rs drops an unused
verify_password import). cargo check + clippy clean, 639 tests green.

Bump 5.97.13 -> 5.97.14.
